starlark: Allow workflow = None to ignore the workflow run
This change lets us ignore a starlark workflow, and fall back to the
standard ci.yaml. This will permit us to provide a "default workflow",
and have that have the option of actively ignoring repositories.

It also adds an optional argument to `loadFile`, `error_on_notfound=True`.
When set to `False`, loading a file that does not exist will not stop
execution and will return None. (note that module `load()` will always
error).

This also lets `getWorkflow` report the user `print`ed content back to
the user via the `debug` field of the API response.